Moved DB.Document extension methods to External.DB.Extensions namespace.
