+++ src/RhinoInside.Revit.GH/Types/Document.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using System.Linq;
 using Grasshopper.Kernel.Types;
+using RhinoInside.Revit.External.DB.Extensions;
 using DB = Autodesk.Revit.DB;
 
 namespace RhinoInside.Revit.GH.Types
