+++ src/RhinoInside.Revit/External/UI/Extensions/UIApplication.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using Microsoft.Win32.SafeHandles;
+using RhinoInside.Revit.External.DB.Extensions;
 
 namespace RhinoInside.Revit.External.UI.Extensions
 {
