+++ src/RhinoInside.Revit/RhinoInside.Revit.csproj
@@ -172,6 +172,7 @@
     <Compile Include="Extensions\RevitAPI.cs" />
     <Compile Include="External\DB\Extensions\Category.cs" />
     <Compile Include="External\DB\Extensions\Curve.cs" />
+    <Compile Include="External\DB\Extensions\Document.cs" />
     <Compile Include="External\DB\Extensions\Element.cs" />
     <Compile Include="External\DB\Extensions\ElementId.cs" />
     <Compile Include="External\DB\Extensions\Face.cs" />
